Prevent duplicate instances and add IPC for check-now command
  - Use setproctitle and psutil to detect existing update-station processes
  - Send SIGUSR1 to running instance when 'check-now' is invoked, triggering update check in existing GTK loop
  - Consolidate notification.py into frontend.py to reduce module count
  - Simplify boolean returns in backend.py
  - Exit with error if tray mode started while instance is already running

This prevents multiple update-station processes from running simultaneously
and allows the check-now command to communicate with the existing tray
instance instead of spawning a separate GUI process.
