--- po/update-station.pot
@@ -1,255 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2026-02-07 12:41-0400\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=CHARSET\n"
-"Content-Transfer-Encoding: 8bit\n"
-
-#: update_station/dialog.py:56
-msgid "Update Failed"
-msgstr ""
-
-#: update_station/dialog.py:61
-msgid ""
-"Press \"Detail\" to get information about the failure.\n"
-"        Get help at https://forums.ghostbsd.org or ."
-msgstr ""
-
-#: update_station/dialog.py:69
-msgid "Detail"
-msgstr ""
-
-#: update_station/dialog.py:71 update_station/dialog.py:126
-#: update_station/dialog.py:157 update_station/dialog.py:187
-#: update_station/dialog.py:218 update_station/dialog.py:249
-#: update_station/dialog.py:282 update_station/dialog.py:308
-#: update_station/frontend.py:118 update_station/notification.py:87
-msgid "Close"
-msgstr ""
-
-#: update_station/dialog.py:87 update_station/dialog.py:116
-msgid "Update Completed"
-msgstr ""
-
-#: update_station/dialog.py:91
-msgid "The computer needs to restart to run on the updated software."
-msgstr ""
-
-#: update_station/dialog.py:98
-msgid "Restart Now"
-msgstr ""
-
-#: update_station/dialog.py:100
-msgid "Restart Later"
-msgstr ""
-
-#: update_station/dialog.py:120
-msgid "All software on this system is up to date."
-msgstr ""
-
-#: update_station/dialog.py:141
-msgid "No Update Available"
-msgstr ""
-
-#: update_station/dialog.py:150
-msgid "No update available. This system is up to date."
-msgstr ""
-
-#: update_station/dialog.py:172
-msgid "Update Station already started"
-msgstr ""
-
-#: update_station/dialog.py:181
-msgid "Update Station already open."
-msgstr ""
-
-#: update_station/dialog.py:202 update_station/dialog.py:233
-msgid "Server Unreachable"
-msgstr ""
-
-#: update_station/dialog.py:211
-msgid "Packages mirrors are syncing with new packages"
-msgstr ""
-
-#: update_station/dialog.py:242
-msgid ""
-"The server is unreachable. Your internet could\n"
-"be down or software package server is down."
-msgstr ""
-
-#: update_station/dialog.py:264
-msgid "Something Is Wrong"
-msgstr ""
-
-#: update_station/dialog.py:274
-msgid ""
-"If you see this message it means that something is wrong.\n"
-" Please look at pkg upgrade output."
-msgstr ""
-
-#: update_station/dialog.py:297
-msgid "Software Station"
-msgstr ""
-
-#: update_station/dialog.py:302
-msgid "You need to be root"
-msgstr ""
-
-#: update_station/frontend.py:105
-msgid "Create boot environment backup"
-msgstr ""
-
-#: update_station/frontend.py:122
-msgid "Install update"
-msgstr ""
-
-#: update_station/frontend.py:135
-msgid "Update Manager"
-msgstr ""
-
-#: update_station/frontend.py:147
-msgid "Updates available!"
-msgstr ""
-
-#: update_station/frontend.py:185
-msgid "Installed packages to be upgraded:"
-msgstr ""
-
-#: update_station/frontend.py:191
-msgid "Installed packages to be downgraded:"
-msgstr ""
-
-#: update_station/frontend.py:197
-msgid "New packages to be installed:"
-msgstr ""
-
-#: update_station/frontend.py:203
-msgid "Installed packages to be reinstalled:"
-msgstr ""
-
-#: update_station/frontend.py:209
-msgid "Installed packages to be removed:"
-msgstr ""
-
-#: update_station/frontend.py:248
-msgid "Installing Update"
-msgstr ""
-
-#: update_station/frontend.py:371
-msgid "Package updates downloaded"
-msgstr ""
-
-#: update_station/frontend.py:374
-msgid "Installing package updates"
-msgstr ""
-
-#: update_station/frontend.py:399
-msgid "Removed"
-msgstr ""
-
-#: update_station/frontend.py:401
-msgid "will reinstall after upgrade"
-msgstr ""
-
-#: update_station/frontend.py:408
-msgid "Software packages upgrade completed"
-msgstr ""
-
-#: update_station/frontend.py:416
-msgid "Reinstalling"
-msgstr ""
-
-#: update_station/frontend.py:438 update_station/frontend.py:459
-msgid "Fetching package updates"
-msgstr ""
-
-#: update_station/frontend.py:478
-msgid "Cleaning old boot environment"
-msgstr ""
-
-#: update_station/frontend.py:486
-msgid "Creating boot environment"
-msgstr ""
-
-#: update_station/frontend.py:564
-msgid "Looking For Updates"
-msgstr ""
-
-#: update_station/frontend.py:604
-msgid "Checking if the repository is online"
-msgstr ""
-
-#: update_station/frontend.py:608
-msgid "The repository is online"
-msgstr ""
-
-#: update_station/frontend.py:612
-msgid "The mirror is Syncing"
-msgstr ""
-
-#: update_station/frontend.py:616
-msgid "Updates are already running"
-msgstr ""
-
-#: update_station/frontend.py:620
-msgid "Checking for updates"
-msgstr ""
-
-#: update_station/frontend.py:623
-msgid "Getting the list of packages"
-msgstr ""
-
-#: update_station/frontend.py:627
-msgid "Open the update window"
-msgstr ""
-
-#: update_station/frontend.py:631
-msgid "No update found"
-msgstr ""
-
-#: update_station/frontend.py:638
-msgid "The Mirror is unreachable"
-msgstr ""
-
-#: update_station/notification.py:25
-msgid "Software updates are now available."
-msgstr ""
-
-#: update_station/notification.py:33
-msgid "Major system version upgrade is now available."
-msgstr ""
-
-#: update_station/notification.py:35
-msgid "System and software updates are now available."
-msgstr ""
-
-#: update_station/notification.py:37
-msgid "Update Available"
-msgstr ""
-
-#: update_station/notification.py:85
-msgid "Open Update"
-msgstr ""
-
-#: update_station/notification.py:130
-msgid "Major version upgrade"
-msgstr ""
-
-#: update_station/notification.py:138
-msgid ""
-"\n"
-"\n"
-"If you select No, the upgrade will be skipped until the next boot."
-msgstr ""
